+++ tests/Recyclarr.Cli.Tests/IntegrationTests/CliCommandIntegrationTest.cs
@@ -1,74 +1,11 @@
 using System.IO.Abstractions;
 using Recyclarr.Cli.Console;
 using Recyclarr.Cli.Tests.Reusable;
-using Recyclarr.TestLibrary;
-using Spectre.Console.Cli;
 
 namespace Recyclarr.Cli.Tests.IntegrationTests;
 
 internal sealed class CliCommandIntegrationTest : CliIntegrationFixture
 {
-    [Test]
-    public async Task List_custom_format_radarr_score_sets()
-    {
-        var reposDir = Paths
-            .ReposDirectory.SubDirectory("trash-guides")
-            .SubDirectory("git")
-            .SubDirectory("official");
-
-        var targetDir = reposDir
-            .SubDirectory("docs")
-            .SubDirectory("json")
-            .SubDirectory("radarr")
-            .SubDirectory("cf");
-        Fs.AddFilesFromEmbeddedNamespace(
-            targetDir,
-            typeof(CliCommandIntegrationTest),
-            "Data.radarr.cfs"
-        );
-
-        var exitCode = await CliSetup.Run(
-            Container,
-            ["list", "custom-formats", "radarr", "--score-sets"]
-        );
-
-        exitCode.Should().Be(0);
-    }
-
-    [Test]
-    public async Task List_custom_format_sonarr_score_sets()
-    {
-        var reposDir = Paths
-            .ReposDirectory.SubDirectory("trash-guides")
-            .SubDirectory("git")
-            .SubDirectory("official");
-
-        var targetDir = reposDir
-            .SubDirectory("docs")
-            .SubDirectory("json")
-            .SubDirectory("sonarr")
-            .SubDirectory("cf");
-        Fs.AddFilesFromEmbeddedNamespace(
-            targetDir,
-            typeof(CliCommandIntegrationTest),
-            "Data.sonarr.cfs"
-        );
-
-        var exitCode = await CliSetup.Run(
-            Container,
-            ["list", "custom-formats", "sonarr", "--score-sets"]
-        );
-
-        exitCode.Should().Be(0);
-    }
-
-    [Test]
-    public async Task List_custom_format_score_sets_fails_without_service_type()
-    {
-        var act = () => CliSetup.Run(Container, ["list", "custom-formats", "--score-sets"]);
-        await act.Should().ThrowAsync<CommandRuntimeException>();
-    }
-
     [Test]
     public async Task List_naming_sonarr()
     {